0001 Improve performance when bases have multiples
When there are multiples in the bases, they don't contribute
in any way to the multiples generated by the calculator, but
the calculator will take a lot of time working with them.

This patch removes multiples from the bases before using
them in the calculator, so the run is much faster.